+++ sw/src/math/vec3i.rs
@@ -3,17 +3,17 @@ use std::ops;
 /// A 3-dimensional vector
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Vec3i {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub x: i16,
+    pub y: i16,
+    pub z: i16,
 }
 
 impl Vec3i {
-    pub fn new(x: i32, y: i32, z: i32) -> Self {
+    pub fn new(x: i16, y: i16, z: i16) -> Self {
         Self { x, y, z }
     }
 
-    pub fn magnitude_squared(self) -> i32 {
+    pub fn magnitude_squared(self) -> i16 {
         (self.x * self.x) + (self.y * self.y) + (self.z * self.z)
     }
 }
@@ -42,10 +42,10 @@ impl ops::Sub for Vec3i {
     }
 }
 
-impl ops::Mul<i32> for Vec3i {
+impl ops::Mul<i16> for Vec3i {
     type Output = Self;
 
-    fn mul(self, rhs: i32) -> Self::Output {
+    fn mul(self, rhs: i16) -> Self::Output {
         Self {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -54,7 +54,7 @@ impl ops::Mul<i32> for Vec3i {
     }
 }
 
-impl ops::Mul<Vec3i> for i32 {
+impl ops::Mul<Vec3i> for i16 {
     type Output = Vec3i;
 
     fn mul(self, rhs: Vec3i) -> Self::Output {
@@ -74,8 +74,8 @@ impl ops::SubAssign for Vec3i {
     }
 }
 
-impl ops::MulAssign<i32> for Vec3i {
-    fn mul_assign(&mut self, rhs: i32) {
+impl ops::MulAssign<i16> for Vec3i {
+    fn mul_assign(&mut self, rhs: i16) {
         *self = *self * rhs;
     }
 }
